Add support for underground coordinates
This adds support for negative coordinates by splitting the altitude dimension in half,
allowing for 2^24 meters above and under sea level.
Also specifies that h=0 is sea level. (For high precision applications, I think this will need to be specified more precisely, for example a specific reference ellipsoid or geoid)